Locking to php 7.3.x , Cleaning up a bunch of stuff. Added scripts. Update README.
- Broke many of the commands to their orn RUN for caching.
- Added scripts for building, running, and clearing tested containers/images as ./docker-*.sh files.
- Updated the README.md
- I ran the test scripts: ./docker-run-test-1*04.sh and everything works.
  - However, when I run the ci_sut image, it fails by saying it can't connect to port 3000/3001 respectivly.
  - Not sure why the port mapping isn't working.

+++ supporting_files/run.sh
@@ -6,9 +6,14 @@ if [ -e /etc/php/5.6/apache2/php.ini ]
 then
     sed -ri -e "s/^upload_max_filesize.*/upload_max_filesize = ${PHP_UPLOAD_MAX_FILESIZE}/" \
         -e "s/^post_max_size.*/post_max_size = ${PHP_POST_MAX_SIZE}/" /etc/php/5.6/apache2/php.ini
-else
+elif [ -e /etc/php/7.2/apache2/php.ini ]
+then
     sed -ri -e "s/^upload_max_filesize.*/upload_max_filesize = ${PHP_UPLOAD_MAX_FILESIZE}/" \
         -e "s/^post_max_size.*/post_max_size = ${PHP_POST_MAX_SIZE}/" /etc/php/7.2/apache2/php.ini
+elif [ -e /etc/php/7.3/apache2/php.ini ]
+then
+    sed -ri -e "s/^upload_max_filesize.*/upload_max_filesize = ${PHP_UPLOAD_MAX_FILESIZE}/" \
+        -e "s/^post_max_size.*/post_max_size = ${PHP_POST_MAX_SIZE}/" /etc/php/7.3/apache2/php.ini
 fi
 
 
@@ -40,7 +45,9 @@ else
     chmod -R 770 /var/run/mysqld
 fi
 
-rm /var/run/mysqld/mysqld.sock
+if [ -e /var/run/mysqld/mysqld.sock ]; then
+    rm /var/run/mysqld/mysqld.sock
+fi
 
 sed -i "s/bind-address.*/bind-address = 0.0.0.0/" /etc/mysql/my.cnf
 sed -i "s/user.*/user = www-data/" /etc/mysql/my.cnf
@@ -58,7 +65,7 @@ if [[ ! -d $VOLUME_HOME/mysql ]]; then
         mysql_install_db > /dev/null 2>&1
     fi
 
-    echo "=> Done!"  
+    echo "=> Done!"
     /create_mysql_users.sh
 else
     echo "=> Using an existing volume of MySQL"
